Started refactoring.
