+++ client/api/emails.js
@@ -1,0 +1,8 @@
+export const sendEmail = async (payload, axios) => {
+  try {
+    const response = await axios.post('api/send-email', payload)
+    return response
+  } catch (error) {
+    return error
+  }
+}
