+++ client/store/emails.js
@@ -1,0 +1,8 @@
+import { sendEmail } from '@/api/emails'
+
+export const actions = {
+  async sendEmail(_, payload) {
+    const response = await sendEmail(payload, this.$axios)
+    return response
+  },
+}
