--- client/store/index.js
@@ -1,9 +1,0 @@
-import blog from './modules/blog'
-import leads from './modules/leads'
-import emails from './modules/emails'
-
-export const modules = {
-  blog,
-  leads,
-  emails,
-}
