+++ client/store/leads.js
@@ -1,0 +1,8 @@
+import { createNewLead } from '@/api/leads'
+
+export const actions = {
+  async createNewLead(_, payload) {
+    const data = await createNewLead(payload, this.$axios)
+    return data
+  },
+}
