--- client/store/modules/leads.js
@@ -1,12 +1,0 @@
-export default {
-  actions: {
-    async createNewLead(_, payload) {
-      try {
-        const response = this.$axios.post('api/leads', payload)
-        return response.data
-      } catch (error) {
-        return error
-      }
-    },
-  },
-}
